# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Data:
    properties:
      code:
        $ref: '#/definitions/response.ResCode'
      data: {}
      msg: {}
    type: object
  response.ResCode:
    enum:
    - 20000
    format: int64
    type: integer
    x-enum-varnames:
    - CodeSuccess
...
  schema.CommentCreateRequest:
    properties:
      commenter_ip:
        description: Commenter IP | 评论者IP
        example: 192.168.1.1
        type: string
      content:
        description: Comment content | 评论内容
        example: 很有见地的评论
        maxLength: 1000
        minLength: 1
        type: string
      device_info:
        description: Device info | 设备信息
        example: Chrome/Windows
        type: string
      parent_id:
        description: Parent comment ID | 父评论ID
        example: 1
        type: integer
      post_id:
        description: Post ID | 帖子ID
        example: 1
        type: integer
      reply_to_user_id:
        description: Reply target user ID | 回复目标用户ID
        example: 2
        type: integer
      user_id:
        description: User ID | 用户ID
        example: 1
        type: integer
    required:
    - content
    - post_id
    - user_id
    type: object
  schema.CommentDetailResponse:
    properties:
      commenter_ip:
        description: Commenter IP | 评论者IP
        example: 192.168.1.1
        type: string
      content:
        description: Comment content | 评论内容
        example: 很有见地的评论
        type: string
      created_at:
        description: Creation time | 创建时间
        example: "2024-01-01 00:00:00"
        type: string
      device_info:
        description: Device info | 设备信息
        example: Chrome/Windows
        type: string
      dislike_count:
        description: Dislike count | 点踩数
        example: 1
        type: integer
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
      is_pinned:
        description: Whether pinned | 是否置顶
        example: false
        type: boolean
      is_selected:
        description: Whether selected | 是否精选
        example: true
        type: boolean
      like_count:
        description: Like count | 点赞数
        example: 10
        type: integer
      parent_id:
        description: Parent comment ID | 父评论ID
        example: 1
        type: integer
      post_id:
        description: Post ID | 帖子ID
        example: 1
        type: integer
      post_title:
        description: Post title | 帖子标题
        example: 技术分享帖
        type: string
      reply_to_user_id:
        description: Reply target user ID | 回复目标用户ID
        example: 2
        type: integer
      reply_to_username:
        description: Reply target username | 回复目标用户名
        example: targetuser
        type: string
      updated_at:
        description: Update time | 更新时间
        example: "2024-01-01 00:00:00"
        type: string
      user_id:
        description: User ID | 用户ID
        example: 1
        type: integer
      username:
        description: Username | 用户名
        example: testuser
        type: string
    type: object
...
  schema.CommentListResponse:
    properties:
      list:
        description: Comment list | 评论列表
        items:
          $ref: '#/definitions/schema.CommentListItem'
        type: array
      page:
        description: Current page number | 当前页码
        type: integer
      page_size:
        description: Items per page | 每页数量
        type: integer
      total:
        description: Total count | 总数量
        type: integer
    type: object
...
  schema.CommentSettingsRequest:
    properties:
      keyword_blacklist:
        description: Keyword blacklist (comma-separated) | 关键词黑名单（逗号分隔）
        example: 垃圾,广告,spam
        maxLength: 5000
        type: string
      require_approval:
        description: Whether comment approval is required | 是否需要审核评论
        example: false
        type: boolean
      show_comment_info:
        description: Whether to show commenter information | 是否显示评论者信息
        example: true
        type: boolean
    type: object
  schema.CommentSettingsResponse:
    properties:
      keyword_blacklist:
        description: Keyword blacklist | 关键词黑名单
        example: 垃圾,广告,spam
        type: string
      require_approval:
        description: Whether comment approval is required | 是否需要审核评论
        example: false
        type: boolean
      show_comment_info:
        description: Whether to show commenter information | 是否显示评论者信息
        example: true
        type: boolean
    type: object
...
  schema.CommentUpdateRequest:
    properties:
      content:
        description: Comment content | 评论内容
        example: 更新后的评论内容
        maxLength: 1000
        minLength: 1
        type: string
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
    required:
    - content
    - id
    type: object
...
  schema.UserCommentActionRequest:
    properties:
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
    required:
    - id
    type: object
  schema.UserCommentActionResponse:
    properties:
      dislike_count:
        description: Dislike count | 点踩数
        example: 1
        type: integer
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
      like_count:
        description: Like count | 点赞数
        example: 10
        type: integer
    type: object
  schema.UserCommentCreateRequest:
    properties:
      content:
        description: Comment content | 评论内容
        example: 很有见地的评论
        minLength: 1
        type: string
      parent_id:
        description: Parent comment ID (used when replying to comment) | 父评论ID（回复评论时使用）
        example: 1
        type: integer
      post_id:
        description: Post ID | 帖子ID
        example: 1
        type: integer
      reply_to_user_id:
        description: Reply target user ID (used when replying to user) | 回复目标用户ID（回复用户时使用）
        example: 2
        type: integer
    required:
    - content
    - post_id
    type: object
  schema.UserCommentCreateResponse:
    properties:
      content:
        description: Comment content | 评论内容
        example: 很有见地的评论
        type: string
      created_at:
        description: Creation time | 创建时间
        example: "2024-01-01 00:00:00"
        type: string
      dislike_count:
        description: Dislike count | 点踩数
        example: 0
        type: integer
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
      like_count:
        description: Like count | 点赞数
        example: 0
        type: integer
      parent_id:
        description: Parent comment ID | 父评论ID
        example: 1
        type: integer
      post_id:
        description: Post ID | 帖子ID
        example: 1
        type: integer
      reply_to_user_id:
        description: Reply target user ID | 回复目标用户ID
        example: 2
        type: integer
      reply_to_username:
        description: Reply target username | 回复目标用户名
        example: targetuser
        type: string
    type: object
...
  schema.UserCommentListResponse:
    properties:
      list:
        description: Comment list | 评论列表
        items:
          $ref: '#/definitions/schema.UserCommentListItem'
        type: array
      page:
        description: Current page number | 当前页码
        type: integer
      page_size:
        description: Items per page | 每页数量
        type: integer
      total:
        description: Total count | 总数量
        type: integer
    type: object
  schema.UserCommentUpdateRequest:
    properties:
      content:
        description: Comment content | 评论内容
        example: 更新后的评论内容
        minLength: 1
        type: string
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
    required:
    - content
    - id
    type: object
  schema.UserCommentUpdateResponse:
    properties:
      content:
        description: Comment content | 评论内容
        example: 更新后的评论内容
        type: string
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
      updated_at:
        description: Update time | 更新时间
        example: "2024-01-01 00:00:00"
        type: string
    type: object
...
  schema.UserProfileCommentItem:
    properties:
      content:
        description: Comment content | 评论内容
        example: 很有见地的评论
        type: string
      created_at:
        description: Creation time | 创建时间
        example: "2024-01-01 00:00:00"
        type: string
      dislike_count:
        description: Dislike count | 点踩数
        example: 1
        type: integer
      id:
        description: Comment ID | 评论ID
        example: 1
        type: integer
      like_count:
        description: Like count | 点赞数
        example: 10
        type: integer
      post_id:
        description: Post ID | 帖子ID
        example: 1
        type: integer
      post_title:
        description: Post title | 帖子标题
        example: 我的第一个帖子
        type: string
    type: object
  schema.UserProfileCommentsResponse:
    properties:
      list:
        description: Comment list | 评论列表
        items:
          $ref: '#/definitions/schema.UserProfileCommentItem'
        type: array
      page:
        description: Current page number | 当前页码
        type: integer
      page_size:
        description: Items per page | 每页数量
        type: integer
      total:
        description: Total count | 总数量
        type: integer
    type: object
...
host: localhost:9876
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: PokeForum is a forum application based on Gin framework | PokeForum
    是一个基于 Gin 框架的论坛应用程序
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PokeForum API
  version: "1.0"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Get paginated comment list for specified post with sorting support
        | 分页获取指定帖子的评论列表,支持排序
      parameters:
      - description: Post ID | 帖子ID
        in: query
        name: post_id
        required: true
        type: integer
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: 'Sort field: created_at, like_count | 排序字段:created_at, like_count'
        example: '"created_at"'
        in: query
        name: sort_by
        type: string
      - description: Is descending order | 是否降序
        in: query
        name: sort_desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved successfully | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentListResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get comment list | 获取评论列表
      tags:
      - '[User]Comments | [用户]评论'
    post:
      consumes:
      - application/json
      description: User creates new comment, supports replying to comments and users
        | 用户创建新评论,支持回复评论和回复用户
      parameters:
      - description: Create comment request | 创建评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created successfully | 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentCreateResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Create comment | 创建评论
      tags:
      - '[User]Comments | [用户]评论'
    put:
      consumes:
      - application/json
      description: User updates their own comment content | 用户更新自己的评论内容
      parameters:
      - description: Update comment request | 更新评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCommentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated successfully | 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentUpdateResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "403":
          description: Insufficient permissions | 权限不足
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update comment | 更新评论
      tags:
      - '[User]Comments | [用户]评论'
  /comments/dislike:
    post:
      consumes:
      - application/json
      description: User dislikes comment, one-way operation cannot be cancelled |
        用户点踩评论,单向操作不可取消
      parameters:
      - description: Dislike comment request | 点踩评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCommentActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Disliked successfully | 点踩成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentActionResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Dislike comment | 点踩评论
      tags:
      - '[User]Comments | [用户]评论'
  /comments/like:
    post:
      consumes:
      - application/json
      description: User likes comment, one-way operation cannot be cancelled | 用户点赞评论,单向操作不可取消
      parameters:
      - description: Like comment request | 点赞评论请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.UserCommentActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Liked successfully | 点赞成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserCommentActionResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server internal error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Like comment | 点赞评论
      tags:
      - '[User]Comments | [用户]评论'
...
  /manage/comments:
    get:
      consumes:
      - application/json
      description: Get paginated comment list with multiple filtering conditions |
        分页获取评论列表,支持多种筛选条件
      parameters:
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      - description: Search keyword | 搜索关键词
        example: '"技术"'
        in: query
        name: keyword
        type: string
      - description: Post ID | 帖子ID
        in: query
        name: post_id
        type: integer
      - description: User ID | 用户ID
        in: query
        name: user_id
        type: integer
      - description: Parent comment ID | 父评论ID
        in: query
        name: parent_id
        type: integer
      - description: Is featured comment | 是否精选评论
        in: query
        name: is_selected
        type: boolean
      - description: Is pinned comment | 是否置顶评论
        in: query
        name: is_pinned
        type: boolean
      - description: Reply target user ID | 回复目标用户ID
        in: query
        name: reply_to_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved successfully | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.CommentListResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get comment list | 获取评论列表
      tags:
      - '[Admin]Comment Management | [管理员]评论管理'
    post:
      consumes:
      - application/json
      description: Admin creates new comment | 管理员创建新评论
      parameters:
      - description: Comment information | 评论信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.CommentCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created successfully | 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.CommentDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Create comment | 创建评论
      tags:
      - '[Admin]Comment Management | [管理员]评论管理'
    put:
      consumes:
      - application/json
      description: Update comment content information | 更新评论的内容信息
      parameters:
      - description: Comment information | 评论信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.CommentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated successfully | 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.CommentDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Update comment information | 更新评论信息
      tags:
      - '[Admin]Comment Management | [管理员]评论管理'
  /manage/comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the specified comment | 删除指定的评论
      parameters:
      - description: Comment ID | 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted successfully | 删除成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Delete comment | 删除评论
      tags:
      - '[Admin]Comment Management | [管理员]评论管理'
    get:
      consumes:
      - application/json
      description: Get detailed information of the specified comment | 获取指定评论的详细信息
      parameters:
      - description: Comment ID | 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved successfully | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.CommentDetailResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get comment detail | 获取评论详情
      tags:
      - '[Admin]Comment Management | [管理员]评论管理'
...
  /profile/comments:
    get:
      consumes:
      - application/json
      description: Get comments published by specified user, supports pagination,
        retrieves current logged-in user's comments if user_id not provided | 获取指定用户发布的评论列表，支持分页，不传user_id则获取当前登录用户的评论
      parameters:
      - description: User ID, queries current logged-in user if not provided | 用户ID，不传则查询当前登录用户
        in: query
        name: user_id
        type: integer
      - description: Page number | 页码
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page | 每页数量
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retrieve successful | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserProfileCommentsResponse'
              type: object
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "401":
          description: Unauthorized | 未授权
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Internal server error | 服务器内部错误
          schema:
            $ref: '#/definitions/response.Data'
      summary: Get user comments list | 获取用户评论列表
      tags:
      - '[User]Profile | [用户]个人中心'
...
  /super/manage/settings/comment:
    get:
      consumes:
      - application/json
      description: Get comment related configuration including review, blacklist,
        etc. | 获取评论相关的配置，包括审核、黑名单等
      produces:
      - application/json
      responses:
        "200":
          description: Success | 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Data'
            - properties:
                data:
                  $ref: '#/definitions/schema.CommentSettingsResponse'
              type: object
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: Get comment settings | 获取评论设置
      tags:
      - '[Super Admin]System Settings | [超级管理员]系统设置'
    post:
      consumes:
      - application/json
      description: Update comment related configuration | 更新评论相关的配置
      parameters:
      - description: Comment settings information | 评论设置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schema.CommentSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated successfully | 更新成功
          schema:
            $ref: '#/definitions/response.Data'
        "400":
          description: Invalid request parameters | 请求参数错误
          schema:
            $ref: '#/definitions/response.Data'
        "500":
          description: Server error | 服务器错误
          schema:
            $ref: '#/definitions/response.Data'
      security:
      - Bearer: []
      summary: Update comment settings | 更新评论设置
      tags:
      - '[Super Admin]System Settings | [超级管理员]系统设置'
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"